requirements.txt
streamlit~=1.15.1
Pillow~=9.3.0
diffusers~=0.9.0
torch~=1.13.0
transformers~=4.25.1
accelerate~=0.15.0